% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmedm.R
\name{pmedm}
\alias{pmedm}
\title{P-MEDM Wrapper}
\usage{
pmedm(
  wt,
  cind,
  cg2,
  cg1,
  sg2,
  sg1,
  include_cg0 = TRUE,
  g1 = NULL,
  g2 = NULL,
  tr_iter = NULL,
  tr_tol = NULL
)
}
\arguments{
\item{wt}{sample weights}

\item{cind}{individual constraints}

\item{cg2}{block group (or target zone) constraints}

\item{cg1}{(or aggregate zone) constraints}

\item{sg2}{group (or target zone) constraint standard errors}

\item{sg1}{tract (or aggregate zone) constraint standard errors}

\item{include_cg0}{whether to include Level 0 (PUMA) constraints (default TRUE)}

\item{g1}{an optional vector of aggregate zone IDs}

\item{g2}{an optional vector of target zone IDs}

\item{tr_iter}{maximum number of inner trust region interations (optional)}

\item{tr_tol}{tolerance for stopping the solver (optional)}
}
\value{
A list containing the P-MEDM allocation matrix, PUMS serials, and block group IDs.
}
\description{
Wrapper for `PMEDMrcpp::pmedm_solve`. Adapted from various examples by Nagle
(2013, 2015). See `PMEDMrcpp::??pmedm_solve`.
}
